Refactor circular dependencies
